requirements.txt
discord.py
async-timeout
numpy
os
pandas
